refactor PrepareIstioHTTPRouteForCR to NewHTTPRoute
