+++ internal/istio/httproute.go
@@ -1,0 +1,36 @@
+package istio
+
+import (
+	"fmt"
+
+	istioapiv1beta1 "istio.io/api/networking/v1beta1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
+)
+
+func NewHTTPRoute(watcher *v1beta2.Watcher) *istioapiv1beta1.HTTPRoute {
+	return &istioapiv1beta1.HTTPRoute{
+		Name: client.ObjectKeyFromObject(watcher).String(),
+		Match: []*istioapiv1beta1.HTTPMatchRequest{
+			{
+				Uri: &istioapiv1beta1.StringMatch{
+					MatchType: &istioapiv1beta1.StringMatch_Prefix{
+						//nolint:nosnakecase // external type
+						Prefix: fmt.Sprintf(prefixFormat, contractVersion, watcher.GetModuleName()),
+					},
+				},
+			},
+		},
+		Route: []*istioapiv1beta1.HTTPRouteDestination{
+			{
+				Destination: &istioapiv1beta1.Destination{
+					Host: destinationHost(watcher.Spec.ServiceInfo.Name, watcher.Spec.ServiceInfo.Namespace),
+					Port: &istioapiv1beta1.PortSelector{
+						Number: uint32(watcher.Spec.ServiceInfo.Port),
+					},
+				},
+			},
+		},
+	}
+}
